refactor(#110): Create interface for Services and replace static methods with instance methods
closes #110

BREAKING CHANGE: 💥 in sct-commons:
  The services which provide static methods are moved from scl to commons package.
  Now  those services implements interfaces, no static methods.
  Expect for PrivateService which is an UtilityClass is renamed to PrivateUtils and moved from scl to util package.


refactor(#110): give PrivateService the new name PrivateUtils and change the package from scl to util.

Changing the name of PrivateService to PrivateUtils and move it from scl to util package , closes #110

refactor(#110): update SclAutomationService


refactor(#110): check review part1

use @RequiredArgsConstructor, instead of @NoArgsConstructor.
use Camel case for PrivateLinkedToStds


refactor(#110): check review part2 [use mockito properly]

Remove unused test package.
Use latest version of mockito.
Fix related to version 5.4.0


refactor(#110): review part3 [naming] remove prefix 'I' and add suffix 'Editor'

The interfaces are moved from scl to api package.

add package doc


refactor(#110): review part4 [seperate SclService features ]

SclEditor (interface) spec are reduced by adding ReadOnly and Write One


refactor(#110): sonar review


docs(#110): update demo project and quickstart page
