+++ docs/examples/example-app/pom.xml
@@ -18,9 +18,9 @@ SPDX-License-Identifier: Apache-2.0
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <maven.compiler.release>11</maven.compiler.release>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+    <maven.compiler.release>17</maven.compiler.release>
   </properties>
 
   <distributionManagement>
@@ -44,13 +44,13 @@ SPDX-License-Identifier: Apache-2.0
     <dependency>
       <groupId>org.lfenergy.compas</groupId>
       <artifactId>sct-commons</artifactId>
-      <version>0.1.0</version>
+      <version>0.2.12</version>
     </dependency>
     <!-- SclAutomationService -->
     <dependency>
       <groupId>org.lfenergy.compas</groupId>
       <artifactId>sct-app</artifactId>
-      <version>0.1.0</version>
+      <version>0.2.12</version>
     </dependency>
     <dependency>
       <groupId>org.junit.jupiter</groupId>
