+++ pom.xml
@@ -44,6 +44,7 @@
         <logback-classic.version>1.4.5</logback-classic.version>
         <assertj.version>3.22.0</assertj.version>
         <lombok.version>1.18.24</lombok.version>
+        <mockito.version>5.5.0</mockito.version>
         <jackson-databind.version>2.13.4.1</jackson-databind.version>
     </properties>
 
@@ -76,6 +77,18 @@
                 <groupId>org.junit.jupiter</groupId>
                 <artifactId>junit-jupiter-params</artifactId>
                 <version>${junit-version}</version>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-core</artifactId>
+                <version>${mockito.version}</version>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-junit-jupiter</artifactId>
+                <version>${mockito.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
