+++ sct-app/src/main/java/org/lfenergy/compas/sct/app/SclAutomationService.java
@@ -5,13 +5,14 @@
 package org.lfenergy.compas.sct.app;
 
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import org.lfenergy.compas.scl2007b4.model.SCL;
+import org.lfenergy.compas.sct.commons.api.SclEditor;
+import org.lfenergy.compas.sct.commons.api.SubstationEditor;
 import org.lfenergy.compas.sct.commons.dto.HeaderDTO;
 import org.lfenergy.compas.sct.commons.dto.SubNetworkDTO;
 import org.lfenergy.compas.sct.commons.dto.SubNetworkTypeDTO;
 import org.lfenergy.compas.sct.commons.exception.ScdException;
-import org.lfenergy.compas.sct.commons.scl.SclService;
-import org.lfenergy.compas.sct.commons.scl.SubstationService;
 
 import java.util.*;
 
@@ -24,7 +25,11 @@ import java.util.*;
  *   <li>{@link SclAutomationService#createSCD(SCL, HeaderDTO, List) Adds all elements under the <b>SCL </b> object from given <b>SSD </b> and <b>STD </b> files}
  *  </ul>
  */
+@RequiredArgsConstructor
 public class SclAutomationService {
+
+    private final SclEditor sclEditor;
+    private final SubstationEditor substationEditor;
 
     /**
      * Possible Subnetwork and ConnectedAP names which should be used in generated SCD in order a have global coherence
@@ -33,10 +38,6 @@ public class SclAutomationService {
     public static final List<SubNetworkTypeDTO> SUB_NETWORK_TYPES = List.of(
             new SubNetworkTypeDTO("RSPACE_PROCESS_NETWORK", SubNetworkDTO.SubnetworkType.MMS.toString(), List.of("PROCESS_AP", "TOTO_AP_GE")),
             new SubNetworkTypeDTO("RSPACE_ADMIN_NETWORK", SubNetworkDTO.SubnetworkType.IP.toString(), List.of("ADMIN_AP", "TATA_AP_EFFACEC")));
-
-    private SclAutomationService() {
-        throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
-    }
 
     /**
      * Create an SCD file from specified parameters, it calls all functions defined in the process one by one, every step
@@ -47,15 +48,15 @@ public class SclAutomationService {
      * @return an SCD object
      * @throws ScdException
      */
-    public static SCL createSCD(@NonNull SCL ssd, @NonNull HeaderDTO headerDTO, List<SCL> stds) throws ScdException {
-        SCL scd = SclService.initScl(headerDTO.getId(), headerDTO.getVersion(), headerDTO.getRevision());
+    public SCL createSCD(@NonNull SCL ssd, @NonNull HeaderDTO headerDTO, List<SCL> stds) throws ScdException {
+        SCL scd = sclEditor.initScl(headerDTO.getId(), headerDTO.getVersion(), headerDTO.getRevision());
         if (!headerDTO.getHistoryItems().isEmpty()) {
             HeaderDTO.HistoryItem hItem = headerDTO.getHistoryItems().get(0);
-            SclService.addHistoryItem(scd, hItem.getWho(), hItem.getWhat(), hItem.getWhy());
+            sclEditor.addHistoryItem(scd, hItem.getWho(), hItem.getWhat(), hItem.getWhy());
         }
-        SubstationService.addSubstation(scd, ssd);
-        SclService.importSTDElementsInSCD(scd, stds, SUB_NETWORK_TYPES);
-        SclService.removeAllControlBlocksAndDatasetsAndExtRefSrcBindings(scd);
+        substationEditor.addSubstation(scd, ssd);
+        sclEditor.importSTDElementsInSCD(scd, stds, SUB_NETWORK_TYPES);
+        sclEditor.removeAllControlBlocksAndDatasetsAndExtRefSrcBindings(scd);
         return scd;
     }
 }
