+++ sct-commons/src/main/java/org/lfenergy/compas/sct/commons/HmiService.java
@@ -2,26 +2,20 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.lfenergy.compas.sct.commons.scl;
+package org.lfenergy.compas.sct.commons;
 
 import org.lfenergy.compas.scl2007b4.model.SCL;
 import org.lfenergy.compas.scl2007b4.model.TFCDA;
+import org.lfenergy.compas.sct.commons.api.HmiEditor;
+import org.lfenergy.compas.sct.commons.scl.SclRootAdapter;
 import org.lfenergy.compas.sct.commons.scl.ied.IEDAdapter;
 
 import java.util.List;
 
-public final class HmiService {
+public class HmiService implements HmiEditor {
 
-    private HmiService() {
-        throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
-    }
-
-    /**
-     * Create the DataSet and ReportControl Blocks for the HMI with the given FCDAs.
-     *
-     * @param fcdas List of FCDA for which we must create the DataSet and ReportControl Blocks
-     */
-    public static void createAllHmiReportControlBlocks(SCL scd, List<TFCDA> fcdas) {
+    @Override
+    public void createAllHmiReportControlBlocks(SCL scd, List<TFCDA> fcdas) {
         SclRootAdapter sclRootAdapter = new SclRootAdapter(scd);
         sclRootAdapter.streamIEDAdapters()
                 .flatMap(IEDAdapter::streamLDeviceAdapters)
