+++ sct-commons/src/main/java/org/lfenergy/compas/sct/commons/api/package-info.java
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2023 RTE FRANCE
+//
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * <h2>commons.api contains interfaces for manipulating SCL Files</h2>
+ */
+package org.lfenergy.compas.sct.commons.api;
